Walljump!!
Had it *working* last night, but the wall jump direction was set to opposite the direction you're facing, but I changed it to be just opposite of the side you're colliding with a wall on. So if you tap left arrow slightly before you want to walljump to the left, it still works.